modificacion de la visualizacion de registros en hexadecimal y modificacion del code.txt para corregir algunos problemas de las funciones

+++ instrucciones.c
@@ -16,6 +16,7 @@
 */
 
 #include "instrucciones.h"
+#include <curses.h>
 /* Macros para asignacion de las banderas */
 #define N 0
 #define Z 1
@@ -286,8 +287,9 @@ uint32_t LDRSH(uint32_t Rn, uint32_t Rm, uint8_t *SRAM)
 
 void STR(uint32_t Rt, uint32_t Rn, uint32_t Rm, uint8_t *SRAM)
 {
-	uint32_t address;
+	uint8_t address;
 	address=Rn+Rm;
+	//mvprintw(1,1,"%X",Rn);
 	SRAM[address]=(uint8_t)Rt; 
 	SRAM[address+1]=((uint8_t)Rt>>8);
 	SRAM[address+2]=((uint8_t)Rt>>16);
@@ -296,7 +298,7 @@ void STR(uint32_t Rt, uint32_t Rn, uint32_t Rm, uint8_t *SRAM)
 
 void STRB(uint32_t Rt, uint32_t Rn, uint32_t Rm, uint8_t *SRAM)
 {
-	uint32_t address;
+	uint8_t address;
 	address=Rn+Rm;
 	SRAM[address]=(uint8_t)Rt;
 }
